Update to use a TVDB instance
Previous versions used a global and hardcoded API key.  This change moves all functions under a TVDB instance that houses the APIKey that is populated when the instance is created.

e.g:
	series, err := GetSeries("The Simpsons")
becomes:
	t := NewTVDB("<apikey>")
	series, err := t.GetSeries("The Simpsons")

This will also allow to keep some state and some other options like default language in the future.